+++ tests/scripts/skip_me.lua
@@ -1,0 +1,2 @@
+-- skip_me.lua: this file should be blacklisted in tests
+skipped = true