+++ tests/scripts/test1.lua
@@ -1,0 +1,3 @@
+-- test1.lua: sets global variable to verify execution
+test1_executed = true
+test1_value = 100